[IAudioCallback] make sure code is compiling after switching from shor* to float* - we still need to check is this correct and if it's not to return back to short

+++ xbmc/cores/paplayer/PAPlayer.h
@@ -179,7 +179,7 @@ private:
   CFileItem*        m_nextFile;
 
   // stuff for visualisation
-  BYTE             m_visBuffer[PACKET_SIZE];
+  float            m_visBuffer[PACKET_SIZE];
   unsigned int     m_visBufferLength;
   IAudioCallback*  m_pCallback;
 };
